Refactor charting components and context management for improved structure and functionality
